spawn: don't leak shared_ptrs on suspended stacks
test_spawn_timer_destruct exposes a memory leak when a coro_handler is
created but never completes. the spawned callee stack cannot hold
shared_ptr references to itself while suspended, or their continuation
will never destruct and unwind the stack

the change splits continuation_context so that the 'caller' context is
owned by spawn_data, and the 'callee' context is allocated in a separate
shared_ptr. the callee's stack holds a shared_ptr to its spawn_data
for access to its handler and caller context on resume.  as long as the
callee stack never holds a shared_ptr to its own context while
suspended, coro_handler cancelation can destroy the callee's context,
which unwinds its stack along with the shared_ptr<spawn_data>

spawn_data is not used with weak_ptr, so can be allocated with
make_shared

note that the c++03 version with fcontext still leaks, because it
doesn't unwind the callee stack on destruction!

Signed-off-by: Casey Bodley <cbodley@redhat.com>